# specs/003-chapter-personalization/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Chapter Personalization API
  description: |
    API for managing user-specific chapter personalizations (bookmarks) in the Humanoid Robotics Book.

    ## Authentication
    All endpoints require a valid Better-Auth session. Session is validated via HTTP-only cookies.

    ## Idempotency
    - POST /api/personalization/chapters: Idempotent (duplicate adds are safe)
    - DELETE /api/personalization/chapters/{chapterPath}: Idempotent (duplicate removes return 204)

    ## Rate Limiting
    - 100 requests per minute per user
    - 429 Too Many Requests returned when limit exceeded
  version: 1.0.0
  contact:
    name: Humanoid Robotics Book Team
    url: https://github.com/your-repo/humanoid-robotics-book

servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://your-production-domain.com
    description: Production server

tags:
  - name: personalization
    description: Operations for managing personalized chapters

paths:
  /api/personalization/chapters:
    get:
      tags:
        - personalization
      summary: Get all personalized chapters for current user
      description: |
        Fetches a list of all chapters that the authenticated user has personalized (bookmarked).
        Results are ordered by creation date (most recent first).
      operationId: getPersonalizedChapters
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Successfully retrieved personalized chapters
          content:
            application/json:
              schema:
                type: object
                properties:
                  chapters:
                    type: array
                    items:
                      $ref: '#/components/schemas/PersonalizedChapter'
                  count:
                    type: integer
                    description: Total number of personalized chapters
                    example: 5
              examples:
                withChapters:
                  summary: User with personalized chapters
                  value:
                    chapters:
                      - id: 1
                        user_id: "user_abc123"
                        chapter_path: "/docs/intro/getting-started"
                        chapter_title: "Getting Started with Humanoid Robotics"
                        chapter_excerpt: "Learn the basics of humanoid robotics and how..."
                        created_at: "2025-12-17T10:30:00Z"
                      - id: 2
                        user_id: "user_abc123"
                        chapter_path: "/docs/basics/ros-fundamentals"
                        chapter_title: "ROS 2 Fundamentals"
                        chapter_excerpt: "Understand the core concepts of ROS 2 for robo..."
                        created_at: "2025-12-17T09:15:00Z"
                    count: 2
                empty:
                  summary: User with no personalizations
                  value:
                    chapters: []
                    count: 0
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - personalization
      summary: Add a chapter to user's personalizations
      description: |
        Adds a chapter to the authenticated user's personalized chapters list.

        **Idempotent**: If the chapter is already personalized, this operation succeeds with no changes.

        The chapter path must be a valid URL path from the Docusaurus site.
      operationId: addPersonalizedChapter
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddPersonalizationRequest'
            examples:
              basic:
                summary: Add chapter with minimal data
                value:
                  chapter_path: "/docs/intro/getting-started"
              complete:
                summary: Add chapter with title and excerpt
                value:
                  chapter_path: "/docs/advanced/motion-planning"
                  chapter_title: "Advanced Motion Planning"
                  chapter_excerpt: "Explore sophisticated motion planning algorithm..."
      responses:
        '201':
          description: Chapter successfully added to personalizations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonalizedChapter'
        '200':
          description: Chapter was already personalized (idempotent operation)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonalizedChapter'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/personalization/chapters/{chapterPath}:
    delete:
      tags:
        - personalization
      summary: Remove a chapter from user's personalizations
      description: |
        Removes a chapter from the authenticated user's personalized chapters list.

        **Idempotent**: If the chapter is not personalized, this operation succeeds with no changes.

        The chapter path should be URL-encoded if it contains special characters.
      operationId: removePersonalizedChapter
      security:
        - sessionAuth: []
      parameters:
        - name: chapterPath
          in: path
          required: true
          description: URL path of the chapter to remove (URL-encoded)
          schema:
            type: string
            example: "%2Fdocs%2Fintro%2Fgetting-started"
          examples:
            encoded:
              summary: URL-encoded path
              value: "%2Fdocs%2Fintro%2Fgetting-started"
            simple:
              summary: Simple path (if server accepts unencoded)
              value: "/docs/intro/getting-started"
      responses:
        '204':
          description: Chapter successfully removed from personalizations (or was not personalized)
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
      description: Better-Auth session cookie (HTTP-only)

  schemas:
    PersonalizedChapter:
      type: object
      required:
        - id
        - user_id
        - chapter_path
        - created_at
      properties:
        id:
          type: integer
          description: Unique identifier for this personalization record
          example: 1
        user_id:
          type: string
          description: User ID from Better-Auth
          example: "user_abc123"
        chapter_path:
          type: string
          description: Full URL path of the personalized chapter
          pattern: ^/[a-zA-Z0-9/_-]+$
          minLength: 1
          maxLength: 500
          example: "/docs/intro/getting-started"
        chapter_title:
          type: string
          nullable: true
          description: Title of the chapter (optional, for display)
          maxLength: 500
          example: "Getting Started with Humanoid Robotics"
        chapter_excerpt:
          type: string
          nullable: true
          description: First 50 characters of chapter content (optional, for preview)
          maxLength: 50
          example: "Learn the basics of humanoid robotics and how..."
        created_at:
          type: string
          format: date-time
          description: Timestamp when the chapter was personalized
          example: "2025-12-17T10:30:00Z"

    AddPersonalizationRequest:
      type: object
      required:
        - chapter_path
      properties:
        chapter_path:
          type: string
          description: Full URL path of the chapter to personalize
          pattern: ^/[a-zA-Z0-9/_-]+$
          minLength: 1
          maxLength: 500
          example: "/docs/intro/getting-started"
        chapter_title:
          type: string
          description: Title of the chapter (optional, extracted from page metadata)
          maxLength: 500
          example: "Getting Started with Humanoid Robotics"
        chapter_excerpt:
          type: string
          description: Preview text for the chapter (optional, first 50 chars of content)
          maxLength: 50
          example: "Learn the basics of humanoid robotics and how..."

    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code or type
          example: "UNAUTHORIZED"
        message:
          type: string
          description: Human-readable error message
          example: "You must be logged in to access this resource"
        details:
          type: object
          description: Additional error details (optional)
          additionalProperties: true

    ValidationErrorResponse:
      type: object
      required:
        - error
        - message
        - validation_errors
      properties:
        error:
          type: string
          example: "VALIDATION_ERROR"
        message:
          type: string
          example: "Request validation failed"
        validation_errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: "chapter_path"
              message:
                type: string
                example: "Invalid chapter path format"

  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "BAD_REQUEST"
            message: "Invalid chapter path format"

    Unauthorized:
      description: Authentication required (no valid session)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            noSession:
              summary: No session cookie present
              value:
                error: "UNAUTHORIZED"
                message: "You must be logged in to access this resource"
            expiredSession:
              summary: Session expired
              value:
                error: "SESSION_EXPIRED"
                message: "Your session has expired. Please log in again."

    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
          example:
            error: "VALIDATION_ERROR"
            message: "Request validation failed"
            validation_errors:
              - field: "chapter_path"
                message: "Chapter path must start with /"
              - field: "chapter_excerpt"
                message: "Chapter excerpt must be 50 characters or less"

    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "RATE_LIMIT_EXCEEDED"
            message: "Too many requests. Please try again later."
      headers:
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
            example: 60

    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "INTERNAL_SERVER_ERROR"
            message: "An unexpected error occurred. Please try again later."

  examples:
    sampleChapter:
      value:
        id: 1
        user_id: "user_abc123"
        chapter_path: "/docs/intro/getting-started"
        chapter_title: "Getting Started with Humanoid Robotics"
        chapter_excerpt: "Learn the basics of humanoid robotics and how..."
        created_at: "2025-12-17T10:30:00Z"